absolute continent sizes
continent size is set independent from map size and ocean size. results
in less of a spectrum of continent sizes, the only sizes being the set
size and a smattering of islands of 1-10 hexes. can't decide if this is
better or worse than having the whole gamut of sizes on any given map