continent size theory
continent size abstraction layer: calculate the ideal list of sizes and
store the list. pick randomly from the list, but remove from the list
based on the actually grown continent's size by finding the closest
match from the list. when the list runs out, tries to fill remaining
area with as few continents as possible.